more lib/provable restructure
