+++ mina-transaction/derived-leaves.ts
@@ -16,7 +16,7 @@ import {
 } from '../lib/binable.js';
 import { Base58, fieldEncodings } from '../../lib/util/base58.js';
 import { dataAsHash } from '../../lib/mina/events.js';
-import { HashHelpers } from '../../lib/provable/hash-generic.js';
+import { HashHelpers } from '../../lib/provable/crypto/hash-generic.js';
 import { prefixes } from '../crypto/constants.js';
 
 export { derivedLeafTypes, derivedLeafTypesSignable, tokenSymbolLength };
